Improved code efficiency
Now there is a clear structure for logging and proper dictionary usage. Furthermore, the time complexity has improved to have a worst case scenario of O(U) where U is the number of users fetched by admins in the list, plus the O(Tp+Ts) where Tp is personal task of the admin user, and shared task of said admin. This is still not optimal for me.